# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact: {}
  description: This is a data wallet server.
  title: Data-Wallet API
  version: "1.0"
paths:
  /domain/bind:
    post:
      consumes:
      - application/json
      description: DomainBind
      parameters:
      - description: domain
        in: body
        name: domain
        required: true
        schema:
          type: string
      - description: address
        in: body
        name: address
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: DomainBind
      tags:
      - domain
  /domain/renewal:
    post:
      consumes:
      - application/json
      description: DomainRenewal
      parameters:
      - description: domain
        in: body
        name: domain
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: DomainRenewal
      tags:
      - domain
  /mfile/download:
    get:
      consumes:
      - application/json
      description: Download
      parameters:
      - description: mdid
        in: query
        name: mdid
        required: true
        type: string
      - description: address
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Download
      tags:
      - mfile
  /mfile/upload/confirm:
    post:
      consumes:
      - application/json
      description: UploadConfirm
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: UploadConfirm
      tags:
      - mfile
  /mfile/upload/create:
    post:
      consumes:
      - application/json
      description: UploadCreate
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: UploadCreate
      tags:
      - mfile
swagger: "2.0"